+++ packages/create-app/templates/react-ts-webpack/esbuild.config.ts
@@ -1,11 +1,14 @@
+import { BuildOptions } from 'esbuild'
 import path from 'path'
 
 /**
  * @var {Partial<import('esbuild').BuildOptions>}
  */
-export default {
+const config: BuildOptions = {
   platform: 'node',
   entryPoints: [path.resolve('src/main/main.ts')],
   bundle: true,
-  target: 'node14.16.0', // electron version target
+  target: 'node16.5.0', // electron version target
 }
+
+export default config