+++ packages/create-app/templates/react-ts-webpack/webpack.config.ts
@@ -117,17 +117,9 @@ const createConfig = (): Configuration => {
     }
   } else {
     configuration.plugins.push(
-      new webpack.HotModuleReplacementPlugin(),
       new ForkTsCheckerWebpackPlugin(),
       new ReactRefreshWebpackPlugin(),
     )
-
-    configuration.devServer = {
-      host: 'localhost',
-      port: 9080,
-      hot: true,
-      overlay: true,
-    }
 
     const entries = configuration.entry as Record<'renderer', string[]>