+++ packages/create-app/templates/react-vite-ts/src/renderer/vite.config.ts
@@ -5,6 +5,6 @@ import { defineConfig } from 'vite'
 export default defineConfig({
   plugins: [reactRefresh()],
   build: {
-    target: 'chrome89',
+    target: 'chrome94', // electron version target
   },
 })