+++ packages/create-app/templates/react-vite/src/main/main.js
@@ -14,8 +14,6 @@ async function createWindow() {
     webPreferences: {
       nodeIntegration: true,
       devTools: true,
-      enableRemoteModule: false,
-      contextIsolation: false,
     },
     show: false,
   })