+++ packages/create-app/templates/react/esbuild.renderer.config.js
@@ -7,5 +7,5 @@ export default {
   platform: 'browser',
   entryPoints: [path.resolve('src/renderer/index.jsx')],
   bundle: true,
-  target: 'chrome89', // electron version target
+  target: 'chrome94', // electron version target
 }