+++ packages/create-app/templates/svelte-ts/src/renderer/vite.config.ts
@@ -6,6 +6,6 @@ import sveltePreprocess from 'svelte-preprocess'
 export default defineConfig({
   plugins: [svelte({ preprocess: sveltePreprocess() })],
   build: {
-    target: 'chrome89',
+    target: 'chrome94', // electron version target
   },
 })