+++ packages/create-app/templates/svelte/src/renderer/main.js
@@ -3,7 +3,7 @@ import App from './App.svelte'
 const app = new App({
   target: document.getElementById('app'),
   props: {
-    name: 'toto',
+    name: 'World',
   },
 })