+++ packages/electron-esbuild/src/builder/webpack.ts
@@ -5,6 +5,7 @@
  */
 
 import { Compiler, Configuration } from 'webpack'
+import WebpackDevServer from 'webpack-dev-server'
 
 import type { Item } from '../config/config'
 import { Logger } from '../console'
@@ -15,6 +16,7 @@ const _logger = new Logger('Builder/Webpack')
 export class WebpackBuilder extends BaseBuilder<Configuration> {
   readonly hasInitialBuild = false
 
+  private readonly _rendererServer: WebpackDevServer
   private readonly _compiler!: Compiler
 
   constructor(readonly _config: Item<Configuration>) {
@@ -22,21 +24,40 @@ export class WebpackBuilder extends BaseBuilder<Configuration> {
 
     try {
       const webpack = require('webpack')
-      require('webpack-dev-server') // check that user can use webpack-dev-server
+      const WebpackDevServer = require('webpack-dev-server') // check that user can use webpack-dev-server
       this._compiler = webpack(this._config.config)
+      this._rendererServer = new WebpackDevServer(
+        {
+          port: 9080,
+          hot: true,
+          client: {
+            overlay: true,
+          },
+        },
+        this._compiler,
+      )
     } catch (e) {
-      if (
-        e.message?.includes('MODULE_NOT_FOUND') ||
-        e.message?.includes('Invalid configuration object')
-      ) {
-        _logger.end(
-          'Your webpack configuration is invalid. Message from webpack',
-          e.message,
-        )
+      if (e instanceof Error) {
+        if (
+          e.message?.includes('Invalid configuration object') &&
+          e.message?.includes('ValidationError')
+        ) {
+          _logger.end(
+            'Your webpack configuration is invalid. Message from webpack',
+            e.message,
+          )
+        }
+
+        if (e.message?.includes('MODULE_NOT_FOUND')) {
+          _logger.end(
+            "It looks like you're trying to use webpack but it's not installed, try running `npm i -D webpack webpack-dev-server`",
+          )
+        }
       }
 
       _logger.end(
-        "It looks like you're trying to use webpack but it's not installed, try running `npm i -D webpack webpack-dev-server`",
+        'An unknown error occurred while trying to configure webpack',
+        e,
       )
     }
   }
@@ -62,23 +83,19 @@ export class WebpackBuilder extends BaseBuilder<Configuration> {
     })
   }
 
-  dev(): void {
+  async dev(): Promise<void> {
     if (this._config.isMain) {
       // TODO: webpack main watch
     } else {
-      const WebpackDevServer = require('webpack-dev-server')
-      const rendererServer = new WebpackDevServer(this._compiler, {
-        hot: true,
-        overlay: true,
-      })
-
-      rendererServer.listen(9080, 'localhost', (err: Error | undefined) => {
-        if (err) {
-          _logger.error(this.env, 'error', err)
+      try {
+        await this._rendererServer.start()
+      } catch (e) {
+        if (e instanceof Error) {
+          _logger.end(this.env, 'WDS error', e)
         } else {
-          _logger.log(this.env, ': starting webpack dev server')
+          _logger.log(this.env, 'WDS starting')
         }
-      })
+      }
     }
   }
 }